Add regular tcp/ip controller that does not require avahi
This is a more classic "client/server" setup where a remote device has to connect to the host running gbridge.

Normally, the "server" in IoT is the end-node, and typically the "client" is the powered hub or computer that requests data from the server.

However, this implementation is particularly useful for testing gbridge functionality.