Create a C++ test harness for executing
In order to reliably run system tests between multiple processes, it's important to be able to monitor most aspects of child process such as,

* exit status
* runtime duration
* standard input
* standard output
* standard error

For that purpose, the "ChildProcess" class was created. An integration test was created for it as well, called "ChildProcessTest" that ensures it is able to perform the required duties.

Additionally, I will now need to subclass ChildProcess, as e.g. RemoteProcess, and ensure that it is able to perform similar duties, but on a remote host (in this case Qemu) via ssh.