Fix baud rate setting in controllers/uart.c
As it turns out, the command-line parameter for setting the baud rate (-b 115200) actually does nothing because an int was being implicitly cast to a speed_t.

While that works fine on Darwin (MacOS), it does not work at all in Linux.

In Linux B115200 is actually 0010002(octal) or 4098 (decimal), so a LUT is really required to translate common rates to speed_t values portably.

Also, instead of reading the current value and then manually changing fields, I chose to initialize struct termios in a more reproducible way. Namely with cfmakeraw(3). I chose to set the input and output rates at the same time using cfsetspeed(3) as well.